Updated Home Structure
